move cse pass methods in cse file
